# docs/swagger.yaml
basePath: /
definitions:
...
  domain.PatchProductRequest:
    properties:
      amount_sell:
        type: integer
      name:
        type: string
    type: object
  domain.Product:
    properties:
      active:
        type: integer
      amount_sell:
        type: integer
      category:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  domain.ProductResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.Product'
        type: array
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for MyNofu application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: MyNofu API
  version: "1.0"
paths:
...
  /private/product:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all products from the product_master table
      parameters:
      - description: Category
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ProductResponse'
        "500":
          description: Database error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Get all products
      tags:
      - products
  /private/product/{id}:
    patch:
      consumes:
      - application/json
      description: Update product name and/or amount_sell
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.PatchProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{"status": "success"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body
          schema:
            type: string
        "500":
          description: Database error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Update a product
      tags:
      - products
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"